change freq type to double, add file support

+++ src/HuffChar.kt
@@ -1,7 +1,7 @@
 /**
  * Huffman character, short for a pair of a char. and its frequency
  */
-class HuffChar(val content: Char?, val frequency: Int) : Comparable<HuffChar> {
+class HuffChar(val content: Char?, val frequency: Double) : Comparable<HuffChar> {
 
     override fun compareTo(other: HuffChar): Int {
         return this.frequency.compareTo(other.frequency)
@@ -11,5 +11,5 @@ class HuffChar(val content: Char?, val frequency: Int) : Comparable<HuffChar> {
         return "Char: ${this.content}, Freq: ${this.frequency}"
     }
 
-    constructor(frequency: Int) : this(null, frequency)
+    constructor(frequency: Double) : this(null, frequency)
 }
